// src/util/enums.ts
export enum iconProvider {
  Entypo_IP = 'Entypo',
  Feather_IP = 'Feather',
  Ionicons_IP = 'Ionicons',
  FontAwesome6_IP = 'FontAwesome6',
  FontAwesome_IP = 'FontAwesome',
  FontAwesome5_IP = 'FontAwesome5',
  MaterialCommunityIcons_IP = 'MaterialCommunityIcons',
  MaterialIcons_IP = 'MaterialIcons',
  Octicons_IP = 'Octicons',
  AntDesign_IP = 'AntDesign',
}

export enum categoryType {
  Shopping = 1,
  Food = 2,
  Transportation = 3,
  Entertainment = 4,
  Health = 5,
  Utility = 6,
  Grocery = 7,
}